fix the issue about the deployment
fix the issue about the deployment
