+++ hobbymatcher/src/main/test/java/com/hobbymatcher/test/BaseTest.java
@@ -1,17 +1,21 @@
 package com.hobbymatcher.test;
 
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
+import static junit.framework.TestCase.assertEquals;
+
 /**
- *
  * 配置spring和junit整合，junit启动时加载springIOC容器
- *
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 // 告诉junit spring配置文件的位置
-@ContextConfiguration({ "classpath:spring/spring-dao.xml", "classpath:spring/spring-service.xml"})
+@ContextConfiguration({"classpath:spring/spring-dao.xml", "classpath:spring/spring-service.xml"})
 public class BaseTest {
-
+    @Test
+    public void testQueryUser() {
+        assertEquals(2, 2);
+    }
 }
