+++ hobbymatcher/src/main/test/java/com/hobbymatcher/test/dao/UserDaoTest.java
@@ -1,0 +1,25 @@
+package com.hobbymatcher.test.dao;
+
+import com.hobbymatcher.dao.UserDao;
+import com.hobbymatcher.entity.User;
+import com.hobbymatcher.test.BaseTest;
+import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import java.util.List;
+
+import static junit.framework.TestCase.assertEquals;
+
+public class UserDaoTest extends BaseTest {
+
+    @Autowired
+    private UserDao userDao;
+
+    @Test
+    public void testQueryUser() {
+        List<User> userList = userDao.queryUser();
+        assertEquals(2, userList.size());
+    }
+
+
+}
